requirements-prod.txt
fastapi
uvicorn[standard]
prisma-client-py
aiosqlite
pydantic
python-dotenv
loguru
